frontend/requirements.txt
streamlit
requests

# Dotenv
python-dotenv